Step6 query + refactoring
