+++ src/components/Main.jsx
@@ -4,7 +4,7 @@ import { Route, Routes, Navigate } from "react-router-native";
 import theme from "../theme";
 import AppBar from "./AppBar";
 import RepositoryList from "./RepositoryList";
-import RepositoryPage from "./RepositoryPage";
+import RepositoryPage from "./RepositoryList/RepositoryPage";
 import CreateReview from "./CreateReview";
 import SignIn from "./SignIn";
 import SignUp from "./SignUp";
