+++ src/components/RepositoryList/index.jsx
@@ -1,51 +1,12 @@
 import { useState } from "react";
 import { FlatList, View, StyleSheet } from "react-native";
-import { Button, Menu } from "react-native-paper";
-import Text from "./Text";
-import useRepositories from "../hooks/useRepositories";
+
+import Text from "../Text";
+import useRepositories from "../../hooks/useRepositories";
 import RepositoryItem from "./RepositoryItem";
+import ListOrder from "./ListOrder";
 
 const ItemSeparator = () => <View style={styles.separator} />;
-
-const ListOrder = ({ setOrderBy, setOrderDirection }) => {
-  const [visible, setVisible] = useState(false);
-  const openMenu = () => setVisible(true);
-  const closeMenu = () => setVisible(false);
-
-  const handleSelection = (orderBy, orderDirection) => {
-    setOrderBy(orderBy);
-    setOrderDirection(orderDirection);
-    closeMenu();
-  };
-
-  return (
-    <View style={styles.orderByContainer}>
-      <Menu
-        visible={visible}
-        onDismiss={closeMenu}
-        anchor={
-          <Button icon="sort" onPress={openMenu}>
-            Order by
-          </Button>
-        }
-        mode="elevated"
-      >
-        <Menu.Item
-          onPress={() => handleSelection("CREATED_AT", "DESC")}
-          title="Latest repositories"
-        />
-        <Menu.Item
-          onPress={() => handleSelection("RATING_AVERAGE", "DESC")}
-          title="Highest rated repositories"
-        />
-        <Menu.Item
-          onPress={() => handleSelection("RATING_AVERAGE", "ASC")}
-          title="Lowest rated repositories"
-        />
-      </Menu>
-    </View>
-  );
-};
 
 export const RepositoryListContainer = ({
   repositories,
@@ -101,10 +62,5 @@ const styles = StyleSheet.create({
   container: {
     display: "flex",
     rowGap: 10,
-  },
-  orderByContainer: {
-    paddingTop: 20,
-    flexDirection: "row",
-    justifyContent: "left",
   },
 });
